Remove geth logs from Logger and combine with ReactNativeLogger
