+++ __mocks__/react-native-fs.ts
@@ -31,11 +31,11 @@ export const write = jest.fn()
 export const downloadFile = jest.fn()
 export const uploadFiles = jest.fn()
 export const touch = jest.fn()
-export const MainBundlePath = jest.fn()
-export const CachesDirectoryPath = jest.fn()
-export const DocumentDirectoryPath = jest.fn()
-export const ExternalDirectoryPath = jest.fn()
-export const ExternalStorageDirectoryPath = jest.fn()
-export const TemporaryDirectoryPath = jest.fn()
-export const LibraryDirectoryPath = jest.fn()
-export const PicturesDirectoryPath = jest.fn()
+export const MainBundlePath = '__MAIN_BUNDLE_PATH__'
+export const CachesDirectoryPath = '__CACHES_DIRECTORY_PATH__'
+export const DocumentDirectoryPath = '__DOCUMENT_DIRECTORY_PATH__'
+export const ExternalDirectoryPath = '__EXTERNAL_DIRECTORY_PATH__'
+export const ExternalStorageDirectoryPath = '__EXTERNAL_STORAGE_DIRECTORY_PATH__'
+export const TemporaryDirectoryPath = '__TEMPORARY_DIRECTORY_PATH__'
+export const LibraryDirectoryPath = '__LIBRARY_DIRECTORY_PATH__'
+export const PicturesDirectoryPath = '__PICTURES_DIRECTORY_PATH__'
